refactor(modal_runtime): split 1887-line module into a package (#100)
modal_runtime.py held the Modal image, the volume, llama-server
lifecycle, learnings persistence, and the OSWorld agent loop in one
file. Each Modal entrypoint that imported from here had to skim past
all the unrelated concerns to find what it needed.

Split, behavior-neutral. ``git mv`` preserves blame on the bulk
(``run_osworld_impl`` keeps its history under ``osworld.py``).

modal_runtime/                    (line counts after split)
  __init__.py        48           back-compat re-exports of all 12
                                  public symbols
  image.py           85           Modal image + vol + GEMMA4_MODEL +
                                  GGUF_CONFIGS
  llama.py          116           download_model + start_llama_server
                                  + the no-leak Popen pattern
  learnings.py       52           load_learnings + get_prior_learning
  osworld.py       1673           run_osworld_impl + OSWorld setup
                                  helpers (was 1887; -214 moved out)

Backward compat preserved
- All six external callers (deploy/modal/{modal_web_tasks.py,
  modal_test_gemma4_cua.py, modal_osworld_direct.py},
  benchmarks/{osworld_chrome.py, osworld_multiapp.py,
  visualwebarena.py}) do
  ``from mantis_agent.modal_runtime import image, run_osworld_impl,
  vol`` etc. Those keep working unchanged.

Tests / lint
- pytest tests/ -q -> 370 passed (same as after #105; pure refactor).
- ruff check on the new package: clean.

Out of scope (acceptance criterion still partial)
- ``osworld.py`` is 1673 lines — over the 800-line target. The bulk is
  one 1437-line ``run_osworld_impl`` function. Decomposing it requires
  extracting helpers and threading state through, which is an
  architectural change rather than a mechanical move. Worth its own
  focused PR.
- ``baseten_server.py`` (1535) split is the next PR against this
  issue.

Closes #100 (partial — see "Out of scope" above; baseten_server split
follows.)
