+++ src/mantis_agent/modal_runtime/image.py
@@ -1,0 +1,85 @@
+"""Modal image and volume definitions shared across deployment entrypoints.
+
+Each Modal app (deploy/modal/*.py, benchmarks/osworld_*.py) imports
+``image`` and ``vol`` from here so they all share the same container
+build. ``GEMMA4_MODEL`` and ``GGUF_CONFIGS`` pick which Gemma 4 GGUF
+variant the runtime downloads — set ``GEMMA4_MODEL`` env var to one of
+``E4B`` / ``26B`` / ``31B``.
+"""
+
+from __future__ import annotations
+
+import os
+
+import modal
+
+vol = modal.Volume.from_name("osworld-data", create_if_missing=True)
+
+# Full image with CUDA, llama.cpp, Docker, OSWorld deps
+image = (
+    modal.Image.from_registry(
+        "nvidia/cuda:12.4.0-devel-ubuntu22.04", add_python="3.11"
+    )
+    .apt_install(
+        "git", "build-essential", "cmake", "curl", "wget",
+        # Docker (for OSWorld's Docker provider)
+        "docker.io",
+        # QEMU (fallback if Docker-in-Docker doesn't work)
+        "qemu-system-x86", "qemu-utils",
+        # OSWorld deps
+        "tesseract-ocr", "net-tools",
+    )
+    .run_commands(
+        # Build llama.cpp with CUDA (only A100 arch to speed up build 5x)
+        "git clone --depth 1 https://github.com/ggerganov/llama.cpp /opt/llama.cpp",
+        # Link against CUDA stubs (no GPU at build time, real libcuda available at runtime)
+        "ln -sf /usr/local/cuda/lib64/stubs/libcuda.so /usr/local/cuda/lib64/libcuda.so.1 && "
+        "ldconfig && "
+        "cd /opt/llama.cpp && cmake -B build -DGGML_CUDA=ON "
+        "-DCMAKE_CUDA_ARCHITECTURES='80' -DLLAMA_CURL=OFF "
+        "&& cmake --build build --target llama-server --config Release -j$(nproc) "
+        "&& rm /usr/local/cuda/lib64/libcuda.so.1",
+        # Clone OSWorld
+        "git clone --depth 1 https://github.com/xlang-ai/OSWorld.git /opt/OSWorld",
+    )
+    .pip_install(
+        # Core OSWorld deps (all unpinned to avoid conflicts with CUDA image)
+        "requests", "pillow", "fabric", "gymnasium", "pytz", "opencv-python-headless",
+        "matplotlib", "psutil", "tqdm", "pandas", "flask", "requests-toolbelt",
+        "filelock", "lxml", "cssselect", "xmltodict", "openpyxl", "python-docx",
+        "python-pptx", "pypdf", "rapidfuzz", "pymupdf", "chardet", "playwright",
+        "backoff", "formulas", "pydrive", "odfpy", "openai", "func-timeout",
+        "beautifulsoup4", "dashscope", "google-generativeai", "PyYAML", "mutagen",
+        "easyocr", "borb", "pypdf2", "pdfplumber", "wrapt-timeout-decorator",
+        "tiktoken", "groq", "docker", "python-dotenv", "tldextract", "anthropic",
+        "json-repair", "loguru", "gdown", "httpx", "huggingface-hub",
+        "ImageHash", "scikit-image", "json-minify",
+        "pyacoustid", "librosa", "fastdtw", "pygame",
+    )
+    .run_commands(
+        "playwright install --with-deps chromium || true",
+    )
+    # Ship the local mantis_agent package (prompts + tool helpers + this module).
+    .add_local_python_source("mantis_agent")
+)
+
+
+GEMMA4_MODEL = os.environ.get("GEMMA4_MODEL", "26B")  # "E4B", "26B", or "31B"
+
+GGUF_CONFIGS = {
+    "E4B": {
+        "repo": "ggml-org/gemma-4-E4B-it-GGUF",
+        "model_file": "gemma-4-e4b-it-Q4_K_M.gguf",
+        "mmproj_file": "mmproj-gemma-4-e4b-it-f16.gguf",
+    },
+    "26B": {
+        "repo": "ggml-org/gemma-4-26b-a4b-it-GGUF",
+        "model_file": "gemma-4-26B-A4B-it-Q4_K_M.gguf",
+        "mmproj_file": "mmproj-gemma-4-26B-A4B-it-f16.gguf",
+    },
+    "31B": {
+        "repo": "ggml-org/gemma-4-31b-it-GGUF",
+        "model_file": "gemma-4-31B-it-Q4_K_M.gguf",
+        "mmproj_file": "mmproj-gemma-4-31B-it-f16.gguf",
+    },
+}
